+++ lib/parser/parse-options.js
