+++ lib/parser/parse-timestamp.js
