+++ lib/parser/parser.js
@@ -317,4 +317,5 @@ Parser.prototype = {
   }
 };
 
+console.log('separate file');
 module.exports = Parser;
