+++ lib/parser/parsing-error.js
