+++ lib/parser/settings.js
