+++ lib/vtt.js
@@ -17,7 +17,7 @@
 /* -*- Mode: Java; tab-width: 2; indent-tabs-mode: nil; c-basic-offset: 2 -*- */
 /* vim: set shiftwidth=2 tabstop=2 autoindent cindent expandtab: */
 var document = require('global/document');
-var parseTimeStamp = require('./parse-timestamp.js');
+var parseTimeStamp = require('./parser/parse-timestamp.js');
 
 var TEXTAREA_ELEMENT = document.createElement("textarea");
 
@@ -777,6 +777,6 @@ WebVTT.processCues = function(window, cues, overlay) {
   })();
 };
 
-WebVTT.Parser = require('./parser.js');
+WebVTT.Parser = require('./parser/parser.js');
 
 module.exports = WebVTT;
